fix more crashes and move test where they belong
