+++ Modules/_sqlite/context.h
@@ -9,6 +9,12 @@ extern "C" {
 
 #include "module.h"
 
+/*
+ * UDF/callback context structure.
+ *
+ * In order to ensure that the state pointer always outlives the callback
+ * context, we make sure it owns a reference to the module itself.
+ */
 typedef struct {
     PyObject_HEAD
     PyObject *callable;
@@ -18,12 +24,6 @@ typedef struct {
 
 #define pysqlite_CallbackContext_CAST(op)   ((pysqlite_CallbackContext *)(op))
 
-/* Allocate a UDF/callback context structure. In order to ensure that the state
- * pointer always outlives the callback context, we make sure it owns a
- * reference to the module itself. create_callback_context() is always called
- * from connection methods, so we use the defining class to fetch the module
- * pointer.
- */
 PyObject *
 pysqlite_create_callback_context(pysqlite_state *state, PyObject *callable);
